+++ apps/comments/lib/AppInfo/Application.php
@@ -5,10 +5,9 @@
  */
 namespace OCA\Comments\AppInfo;
 
-use Closure;
 use OCA\Comments\Capabilities;
-use OCA\Comments\EventHandler;
 use OCA\Comments\Listener\CommentsEntityEventListener;
+use OCA\Comments\Listener\CommentsEventListener;
 use OCA\Comments\Listener\LoadAdditionalScripts;
 use OCA\Comments\Listener\LoadSidebarScripts;
 use OCA\Comments\MaxAutoCompleteResultsInitialState;
@@ -22,9 +21,8 @@ use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
 use OCP\Comments\CommentsEntityEvent;
-use OCP\Comments\ICommentsManager;
+use OCP\Comments\CommentsEvent;
 use OCP\ISearch;
-use OCP\IServerContainer;
 
 class Application extends App implements IBootstrap {
 	public const APP_ID = 'comments';
@@ -48,6 +46,11 @@ class Application extends App implements IBootstrap {
 			CommentsEntityEvent::class,
 			CommentsEntityEventListener::class
 		);
+		$context->registerEventListener(
+			CommentsEvent::class,
+			CommentsEventListener::class,
+		);
+
 		$context->registerSearchProvider(CommentsSearchProvider::class);
 
 		$context->registerInitialStateProvider(MaxAutoCompleteResultsInitialState::class);
@@ -56,14 +59,6 @@ class Application extends App implements IBootstrap {
 	}
 
 	public function boot(IBootContext $context): void {
-		$context->injectFn(Closure::fromCallable([$this, 'registerCommentsEventHandler']));
-
 		$context->getServerContainer()->get(ISearch::class)->registerProvider(LegacyProvider::class, ['apps' => ['files']]);
-	}
-
-	protected function registerCommentsEventHandler(IServerContainer $container): void {
-		$container->get(ICommentsManager::class)->registerEventHandler(function (): EventHandler {
-			return $this->getContainer()->get(EventHandler::class);
-		});
 	}
 }
