+++ apps/comments/tests/Unit/EventHandlerTest.php
@@ -6,14 +6,14 @@
 namespace OCA\Comments\Tests\Unit\Notification;
 
 use OCA\Comments\Activity\Listener as ActivityListener;
-use OCA\Comments\EventHandler;
+use OCA\Comments\Listener\CommentsEventListener;
 use OCA\Comments\Notification\Listener as NotificationListener;
 use OCP\Comments\CommentsEvent;
 use OCP\Comments\IComment;
 use Test\TestCase;
 
 class EventHandlerTest extends TestCase {
-	/** @var  EventHandler */
+	/** @var  CommentsEventListener */
 	protected $eventHandler;
 
 	/** @var ActivityListener|\PHPUnit\Framework\MockObject\MockObject */
@@ -33,7 +33,7 @@ class EventHandlerTest extends TestCase {
 			->disableOriginalConstructor()
 			->getMock();
 
-		$this->eventHandler = new EventHandler($this->activityListener, $this->notificationListener);
+		$this->eventHandler = new CommentsEventListener($this->activityListener, $this->notificationListener);
 	}
 
 	public function testNotFiles() {
